Consolidated stratum and getwork within libpoolprotocols
