+++ ethminer/CMakeLists.txt
@@ -14,7 +14,7 @@ add_dependencies(${EXECUTABLE} BuildInfo.h)
 
 target_link_libraries(${EXECUTABLE} ethcore)
 target_link_libraries(${EXECUTABLE} ethash)
-target_link_libraries(${EXECUTABLE} ethstratum devcore libjson-rpc-cpp::client)
+target_link_libraries(${EXECUTABLE} poolprotocols devcore libjson-rpc-cpp::client)
 
 if(ETHDBUS)
 	find_package(PkgConfig)
@@ -28,9 +28,6 @@ endif()
 if(APICORE)
 	target_link_libraries(${EXECUTABLE} apicore)
 endif()
-
-target_link_libraries(${EXECUTABLE} ethstratum)
-target_link_libraries(${EXECUTABLE} ethgetwork)
 
 include(GNUInstallDirs)
 install(TARGETS ethminer DESTINATION ${CMAKE_INSTALL_BINDIR})
