+++ libpoolprotocols/PoolClient.cpp
