+++ libpoolprotocols/PoolManager.cpp
