+++ libpoolprotocols/getwork/EthGetworkClient.cpp
