+++ libpoolprotocols/getwork/getwork .json
