+++ libpoolprotocols/getwork/jsonrpc_getwork.h
