+++ libpoolprotocols/stratum/EthStratumClient.cpp
