+++ libpoolprotocols/stratum/EthStratumClient.h
