+++ libpoolprotocols/stratum/EthStratumClientV2.cpp
