+++ libpoolprotocols/stratum/EthStratumClientV2.h
